requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.5.2
asn1==2.4.1
asn1crypto==1.4.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto3==1.25.0
botocore==1.28.0
bs4==0.0.1
cachetools==4.1.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
configparser==5.0.1
cryptography==38.0.3
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.10
Django==4.1.3
django-ajax-datatable==4.4.4
django-filter==22.1
django-mysql==4.7.1
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.16
enum-compat==0.0.3
enum34==1.1.10
futures==3.1.1
fuzzywuzzy==0.18.0
gunicorn==20.0.4
gitpython==3.1.30
httplib2==0.19.0
hvac==0.10.5
idna==2.10
importlib-metadata==3.1.1
influxdb==5.3.1
ipaddress==1.0.23
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.3
jmespath==0.10.0
Markdown==3.3.3
MarkupSafe==1.1.1
msgpack==1.0.0
ndg-httpsclient==0.5.1
oauth2client==4.1.3
oauthlib==3.1.0
packaging==21.0
paramiko==2.7.2
parso==0.7.1
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
pre-commit==2.21.0
prompt-toolkit==3.0.8
ptpython==3.0.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyflakes==2.5.0
PyGithub==1.56
Pygments==2.13.0
PyJWT==2.6.0
PyMySQL==0.10.1
PyNaCl==1.4.0
pyOpenSSL==22.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-Levenshtein==0.12.0
python3-openid==3.2.0
pytz==2020.4
pyvim==3.0.2
pyvmomi==7.0.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.6.0
scandir==1.10.0
simplegeneric==0.8.1
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.0.1
sqlparse==0.4.3
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.4.0
selenium==4.6.0
pyotp==2.7.0
attrs==21.3.0
pytest==7.2.0
pytest-django==4.5.2
pytest-cov==4.0.0
pytest-mock==3.10.0
pylama==8.4.1
semgrep==1.2.1
mysql-connector-python==8.0.26
